% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneSamplePoisson.Design.R
\name{OneSamplePoisson.Design}
\alias{OneSamplePoisson.Design}
\title{One sample Poisson model - Trial Design}
\usage{
OneSamplePoisson.Design(
  prior,
  nmin = 10,
  nmax = 100,
  m0,
  m1,
  d = 0,
  ps,
  pf,
  power = 0.8,
  t1error = 0.05,
  alternative = c("less", "greater"),
  seed = 202209,
  sim = 1000
)
}
\arguments{
\item{prior}{A list of length 3 containing the distributional information of the prior.
The first element is a number specifying the type of prior. Options are
\enumerate{
\item DIP ;
\item Gamma(a,b), where a = shape, b = rate}
The second and third elements of the list are the parameters a and b, respectively.}

\item{nmin}{The start searching sample size}

\item{nmax}{The stop searching sample size}

\item{m0}{The null event rate, which could be taken as the standard or current event rate.}

\item{m1}{The event rate of the new treatment.}

\item{d}{The target improvement (minimal clinically meaningful difference).}

\item{ps}{The efficacy boundary (upper boundary).}

\item{pf}{The futility boundary (lower boundary).}

\item{power}{The expected power to achieve.}

\item{t1error}{The controlled type-I-error.}

\item{alternative}{less (lower values imply greater efficacy) or greater (larger
values imply greater efficacy).}

\item{seed}{The seed for simulations.}

\item{sim}{The number of simulations.}
}
\value{
A list of the arguments with method and computed elements
}
\description{
Calculate the minimum planned sample size under an admissible design.
The users decide the power and type-I-error, and pick the efficacy and futility boundaries.
If there are no admissible design based on controlled type-I-error, then default to output
the designs with the lowest type-I-error and at least the user-defined (e.g. 80\%) power.
}
\examples{
\donttest{
# with traditional Bayesian prior Gamma(0.5,0.001)
OneSamplePoisson.Design(list(2,0.5,0.001), nmin = 10, nmax=100, m0 = 5, m1 = 4, d = 0,
                   ps = 0.95, pf = 0.05, power = 0.80, t1error=0.05, alternative = "less",
                   seed = 202210, sim = 10)
# with DIP
OneSamplePoisson.Design(list(1,0,0), nmin = 10, nmax=100, m0 = 5, m1 = 4, d = 0,
                   ps = 0.95, pf = 0.05, power = 0.80, t1error=0.05, alternative = "less",
                   seed = 202210, sim = 10)
}
}
